xyz2lmp.sh: can be used inline with "--" instead of a filename

+++ src/xyz2lmp.sh
@@ -3,7 +3,8 @@
 # creates final.xyz, converts it to lmp data file, corrects it and invokes
 # lammps for rdf calculation
 # 
-# syntax: $0 <dump.xyz> <log.lammps>
+# syntax: $0 <dump.xyz>
+# note: use "--" instead of a file name to read from stdin
 
 # list of required software
 required="python atomsk perl"
@@ -11,6 +12,7 @@ required="python atomsk perl"
 # show syntax and exit
 fail(){
   echo "syntax: $0 <dump.xyz>" >&2
+  echo "use '--' instead of a file name to read from stdin" >&2
   exit 1
 }
 
@@ -38,14 +40,22 @@ fi
 # validate input
 (( ${#@} == 1 )) || fail
 xyzfile="$1"
-[ -f "$xyzfile" ] || fail
+if [ "$xyzfile" == "--" ] ; then
+  xyzfile=/tmp/pid$$_$RANDOM.xyz
+  xyzisintmp="yes"
+  cat > $xyzfile
+fi
+[ -f "$xyzfile" ] || [ -z "$xyzfile" ] || fail
 lmpfile=/tmp/pid$$_$RANDOM.lmp
 
 # retrieve unique types in order
-types=`echo $(sed -n '3,/^\s*[0-9]*\s*$/ s/^\s*\([A-Z][a-z]*\)\s.*$/\1/p' $xyzfile | perl -ne 'if (!defined $x{$_}) { print $_; $x{$_} = 1; }' )`
+types=`sed -n '3,/^\s*[0-9]*\s*$/ s/^\s*\([A-Z][a-z]*\)\s.*$/\1/p' "$xyzfile" | perl -ne 'if (!defined $x{$_}) { print $_; $x{$_} = 1; }' | xargs`
 
 rm "$lmpfile" 2>/dev/null
 atomsk $xyzfile "$lmpfile" >/dev/null
+
+# cleanup of temporary files
+[ "$xyzisintmp" == "yes" ] && rm $xyzfile
 
 # add zero charge to every atom if required (might work with reaxff only)
 if [ "$(echo $atomstyle)" == "atom_style charge" ]; then
